ptyhost: move the frame protocol to a build-tag-free frame.go
So the coming Windows ConPTY host (ptyhost_windows.go) can share the
exact wire format with the Unix host and cux attach, without duplicating
it. Pure move — the Unix host is otherwise unchanged (tests still green).
